Tweak the polling rate of SMBus sensors and remove UUID field
The MMC code runs on a slow processor which can easily be overwhelmed
by messages. Reduce the polling rate of SMBus sensors (which are
proxied via the MMC) to try and reduce the number of times the MMC
NACKs requests and causes the sensor reading to error out.

Also, remove the UUID field.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>