meta-ampere/meta-adlink: Update MTD partitions
Update the kernel device tree MTD partition list.
Previously, the 'pnor-uefi' partition actually covered both TF-A (ATF)
and UEFI areas of the SPI-NOR. Fix it so it only covers the UEFI
portion. Add new partitions 'pnor-code' to cover both TF-A and UEFI,
and 'pnor-tfa' to cover just the TF-A area.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>